+++ lib/presets.d.ts
@@ -18,44 +18,30 @@ declare const _default: ({
         seconds: {
             minValue: number;
             maxValue: number;
-            lowerLimit?: undefined;
-            upperLimit?: undefined;
         };
         minutes: {
             minValue: number;
             maxValue: number;
-            lowerLimit?: undefined;
-            upperLimit?: undefined;
         };
         hours: {
             minValue: number;
             maxValue: number;
-            lowerLimit?: undefined;
-            upperLimit?: undefined;
         };
         daysOfMonth: {
             minValue: number;
             maxValue: number;
-            lowerLimit?: undefined;
-            upperLimit?: undefined;
         };
         months: {
             minValue: number;
             maxValue: number;
-            lowerLimit?: undefined;
-            upperLimit?: undefined;
         };
         daysOfWeek: {
             minValue: number;
             maxValue: number;
-            lowerLimit?: undefined;
-            upperLimit?: undefined;
         };
         years: {
             minValue: number;
             maxValue: number;
-            lowerLimit?: undefined;
-            upperLimit?: undefined;
         };
     };
 } | {
@@ -75,108 +61,34 @@ declare const _default: ({
         seconds: {
             minValue: number;
             maxValue: number;
-            lowerLimit?: undefined;
-            upperLimit?: undefined;
         };
         minutes: {
             minValue: number;
             maxValue: number;
-            lowerLimit?: undefined;
-            upperLimit?: undefined;
         };
         hours: {
             minValue: number;
             maxValue: number;
-            lowerLimit?: undefined;
-            upperLimit?: undefined;
         };
         daysOfMonth: {
             minValue: number;
             maxValue: number;
-            lowerLimit?: undefined;
-            upperLimit?: undefined;
         };
         months: {
             minValue: number;
             maxValue: number;
-            lowerLimit?: undefined;
-            upperLimit?: undefined;
         };
         daysOfWeek: {
             minValue: number;
             maxValue: number;
-            lowerLimit?: undefined;
-            upperLimit?: undefined;
         };
         years: {
             minValue: number;
             maxValue: number;
-            lowerLimit?: undefined;
-            upperLimit?: undefined;
         };
         lobaroMustHaveBlankDayField?: undefined;
         lobaroUseHashValue?: undefined;
         lobaroUseListOfNearestWeekdays?: undefined;
-    };
-} | {
-    name: string;
-    preset: {
-        presetId: string;
-        useSeconds: boolean;
-        useYears: boolean;
-        useAliases: boolean;
-        useBlankDay: boolean;
-        allowOnlyOneBlankDayField: boolean;
-        mustHaveBlankDayField: boolean;
-        useLastDayOfMonth: boolean;
-        useLastDayOfWeek: boolean;
-        useNearestWeekday: boolean;
-        lobaroUseListOfNearestWeekdays: boolean;
-        lobaroUseHashValue: boolean;
-        lobaroMustHaveBlankDayField: boolean;
-        useNthWeekdayOfMonth: boolean;
-        seconds: {
-            minValue: number;
-            maxValue: number;
-            lowerLimit: number;
-            upperLimit: number;
-        };
-        minutes: {
-            minValue: number;
-            maxValue: number;
-            lowerLimit: number;
-            upperLimit: number;
-        };
-        hours: {
-            minValue: number;
-            maxValue: number;
-            lowerLimit: number;
-            upperLimit: number;
-        };
-        daysOfMonth: {
-            minValue: number;
-            maxValue: number;
-            lowerLimit: number;
-            upperLimit: number;
-        };
-        months: {
-            minValue: number;
-            maxValue: number;
-            lowerLimit: number;
-            upperLimit: number;
-        };
-        daysOfWeek: {
-            minValue: number;
-            maxValue: number;
-            lowerLimit: number;
-            upperLimit: number;
-        };
-        years: {
-            minValue: number;
-            maxValue: number;
-            lowerLimit: number;
-            upperLimit: number;
-        };
     };
 })[];
 export default _default;