+++ src/presets.ts
@@ -143,58 +143,44 @@ export default [
       presetId: 'lobaro-device-cron',
       useSeconds: true,
       useYears: false,
-      useAliases: false,
+      useAliases: true,
       useBlankDay: true,
       allowOnlyOneBlankDayField: false,
-      mustHaveBlankDayField: true,
+      mustHaveBlankDayField: false,
       useLastDayOfMonth: true,
       useLastDayOfWeek: true,
       useNearestWeekday: true,
       lobaroUseListOfNearestWeekdays: true,
       lobaroUseHashValue: true,
-      lobaroMustHaveBlankDayField: false,
+      lobaroMustHaveBlankDayField: true,
       useNthWeekdayOfMonth: false,
       seconds: {
         minValue: 0,
         maxValue: 59,
-        lowerLimit: 0, // optional, default to minValue
-        upperLimit: 59, // optional, default to maxValue
       },
       minutes: {
         minValue: 0,
         maxValue: 59,
-        lowerLimit: 0, // optional, default to minValue
-        upperLimit: 59, // optional, default to maxValue
       },
       hours: {
         minValue: 0,
         maxValue: 23,
-        lowerLimit: 0, // optional, default to minValue
-        upperLimit: 23, // optional, default to maxValue
       },
       daysOfMonth: {
         minValue: 1,
         maxValue: 31,
-        lowerLimit: 1, // optional, default to minValue
-        upperLimit: 31, // optional, default to maxValue
       },
       months: {
         minValue: 0,
-        maxValue: 11,
-        lowerLimit: 0, // optional, default to minValue
-        upperLimit: 12, // optional, default to maxValue
+        maxValue: 12, // will be converted from 1-12 to 0-11 internally, 0 is also accepted...
       },
       daysOfWeek: {
         minValue: 1,
         maxValue: 7,
-        lowerLimit: 1, // optional, default to minValue
-        upperLimit: 7, // optional, default to maxValue
       },
       years: {
         minValue: 1970,
         maxValue: 2099,
-        lowerLimit: 1970, // optional, default to minValue
-        upperLimit: 2099, // optional, default to maxValue
       },
     },
   },